roles/lobsters: re-enable lobsters-cron by unique naming lobsters-daily.
The name parameter is used by Ansible to distinguish lines in crontab.
With a duplicate name, the previous definition was overwritten.

Restoring this cronjob caused a knock-on problem: the bot received
so many notices at once that it was k-lined on Freenode.  I
temporarily disabled notifications while the new story submissions
cleared, approximately 36 hours worth.

Thank you deedra at Freenode for restoring service.